fix: missing discriminator keyword error message
Improve the error message when discriminator is missing.

Also move some validation tests to test_yaml_model
